+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.4.12-dev1
 
-* Adds console_entrypoint for unstructured-ingest and more structure/docs related to ingest.
+* Adds console_entrypoint for unstructured-ingest, other structure/doc updates related to ingest.
 
 ## 0.4.11